Move singularity engine config package out of internal
Plugins use the code currently living under
internal/pkg/runtime/engines/singularity/config. Since plugins are not
part of singularity's code, they cannot use packages internal to
singularity. Move this package out of internal into
pkg/runtime/engines/singularity/config.

Adjust all the code referring the old path.
