+++ examples/plugins/test-plugin/main.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
-	singularity "github.com/sylabs/singularity/internal/pkg/runtime/engines/singularity/config"
+	singularity "github.com/sylabs/singularity/pkg/runtime/engines/singularity/config"
 	pluginapi "github.com/sylabs/singularity/pkg/plugin"
 )
